Add files via upload
To use on raw case and control data that have a row of data for one year of data. This code duplicates locations and creates new rows for each year of data that we have for each case and control. For instance, a case may have data for residence during year 2000 and another record of data for their second residence in 2010 (only two rows of data). This code will create duplicate rows for years 2001-2009.